[CI-SKIP] Reimplement PlayerManager using ArrayDequeue
Using ArrayDequeue in order to be able to place Players at the front and back of the queue. The class PlayerServerPair (nested class of PlayerManager) is used to store UUIDs and "destination servers". The getPreviousServer method has been replaced with the next method that uses ArrayDequeue.poll to return the next PlayerServerPair in the queue. The addPlayer method has been replaced by the queuePlayer method. The isPlayerRegistered method has been rewritten to work with PlayerServerPair instances.

The checkPriorityAndQueuePlayer is added. It checks for the priority permission and places players that have it at the front of the queue.

CI-SKIP as it will not compile, nothing has been adjusted to the new PlayerManager yet.
Partially implements #10. CC #22. Closes #23.